change all time and canvas arithmetics to double, use textures for blocks

+++ core/src/main/java/net/yudichev/ntetris/ControlState.java
@@ -3,5 +3,5 @@ package net.yudichev.ntetris;
 import java.util.function.Consumer;
 
 public interface ControlState {
-    void forAllPressedKeys(long gameTimeMillis, Consumer<GameControl> activeControlConsumer);
+    void forAllPressedKeys(double gameTimeMillis, Consumer<GameControl> activeControlConsumer);
 }
