+++ core/src/main/java/net/yudichev/ntetris/Game.java
@@ -1,5 +1,5 @@
 package net.yudichev.ntetris;
 
 public interface Game {
-    void render(long gameTimeMillis);
+    void render(double gameTimeMillis);
 }
